Delay removal of packages as much as possible.
Now that OPAM is able to distinguih between the 'build' and 'install'
rules, we want, when reinstalling a package, to build the new version
before to remove the old one: whenever the new version fails to build,
the old one is still installed and the opam repository is left
untouched.

Unfortunatly, we might not apply this strategy with all packages. When a
package A depends on a package B, for preserving the consistency of then
.opam directory we still need to remove A before to remove B. Then, we
can't wait for B to be built before to remove the old version. Still, we
can wait for A to be successfully built.

+++ src/state/opamAction.ml
@@ -145,10 +145,10 @@ let process_dot_install st nv =
 (* Prepare the package build:
    * apply the patches
    * substitute the files *)
-let prepare_package_build st nv =
+let prepare_package_build st nv dir =
   let opam = OpamSwitchState.opam st nv in
 
-  (* Substitute the patched files.*)
+  (* Substitute the patched files. *)
   let patches = OpamFile.OPAM.patches opam in
 
   let iter_patches f =
@@ -173,10 +173,8 @@ let prepare_package_build st nv =
     ignore (iter_patches print_apply)
   else
 
-  let p_build = OpamPath.Switch.build st.switch_global.root st.switch nv in
-
-  OpamFilename.mkdir p_build;
-  OpamFilename.in_dir p_build (fun () ->
+  OpamFilename.mkdir dir;
+  OpamFilename.in_dir dir (fun () ->
     let all = OpamFile.OPAM.substs opam in
     let patches =
       OpamStd.List.filter_map (fun (f,_) ->
@@ -190,17 +188,16 @@ let prepare_package_build st nv =
   (* Apply the patches *)
   let patching_errors =
     iter_patches (fun base ->
-      let root = OpamPath.Switch.build st.switch_global.root st.switch nv in
-      let patch = root // OpamFilename.Base.to_string base in
+      let patch = dir // OpamFilename.Base.to_string base in
       print_apply base;
-      OpamFilename.patch patch p_build)
+      OpamFilename.patch patch dir)
   in
 
   (* Substitute the configuration files. We should be in the right
      directory to get the correct absolute path for the
      substitution files (see [substitute_file] and
      [OpamFilename.of_basename]. *)
-  OpamFilename.in_dir p_build (fun () ->
+  OpamFilename.in_dir dir (fun () ->
     List.iter
       (OpamFilter.expand_interpolations_in_file (OpamPackageVar.resolve ~opam st))
       (OpamFile.OPAM.substs opam)
@@ -208,8 +205,7 @@ let prepare_package_build st nv =
   if patching_errors <> [] then (
     let msg =
       Printf.sprintf "These patches didn't apply at %s:\n%s"
-        (OpamFilename.Dir.to_string
-           (OpamPath.Switch.build st.switch_global.root st.switch nv))
+        (OpamFilename.Dir.to_string dir)
         (OpamStd.Format.itemize (fun x -> x) patching_errors)
     in
     failwith msg
@@ -249,19 +245,17 @@ let download_package st nv =
   in
   OpamProcess.Job.catch (fun e -> Done (`Error (Printexc.to_string e))) job
 
-let extract_package st source nv =
+let extract_package st source nv destdir =
   log "extract_package: %a from %a"
     (slog OpamPackage.to_string) nv
     (slog (OpamStd.Option.to_string OpamTypesBase.string_of_generic_file))
     source;
   if OpamStateConfig.(!r.dryrun) then () else
-  let build_dir = OpamPath.Switch.build st.switch_global.root st.switch nv in
-  OpamFilename.rmdir build_dir;
   let () =
     match source with
     | None -> ()
-    | Some (D dir) -> OpamFilename.copy_dir ~src:dir ~dst:build_dir
-    | Some (F archive) -> OpamFilename.extract archive build_dir
+    | Some (D dir) -> OpamFilename.copy_dir ~src:dir ~dst:destdir
+    | Some (F archive) -> OpamFilename.extract archive destdir
   in
   let is_repackaged_archive =
     Some (F (OpamPath.archive st.switch_global.root nv)) = source
@@ -281,17 +275,17 @@ let extract_package st source nv =
           OpamDownload.download_as
             ~overwrite:true
             ~checksum url
-            (OpamFilename.create build_dir fname))
+            (OpamFilename.create destdir fname))
       (OpamFile.OPAM.extra_sources opam);
 
     List.iter (fun (src,base,hash) ->
         if OpamFilename.digest src <> hash then
           OpamConsole.error_and_exit "Bad hash for %s"
             (OpamFilename.to_string src);
-        OpamFilename.copy ~src ~dst:(OpamFilename.create build_dir base))
+        OpamFilename.copy ~src ~dst:(OpamFilename.create destdir base))
       (OpamFile.OPAM.get_extra_files opam)
   );
-  prepare_package_build st nv
+  prepare_package_build st nv destdir
 
 (* unused ?
 let string_of_commands commands =
@@ -352,7 +346,7 @@ let cmd_wrapper t opam getter cmd args =
 
 (* Remove a given package *)
 let remove_package_aux
-    t ?(keep_build=false) ?(silent=false) ?changes ?force nv =
+    t ?(silent=false) ?changes ?force nv =
   log "Removing %a" (slog OpamPackage.to_string) nv;
   let name = nv.name in
 
@@ -480,9 +474,6 @@ let remove_package_aux
     OpamFilename.remove
       (OpamFile.filename
          (OpamPath.Switch.config t.switch_global.root t.switch name));
-    if not (keep_build || OpamStateConfig.(!r.keep_build_dir)) then
-      OpamFilename.rmdir
-        (OpamPath.Switch.build t.switch_global.root t.switch nv);
     uninstall_files ();
     OpamFilename.remove (OpamFile.filename dot_install);
     revert_changes ();
@@ -503,9 +494,9 @@ let cleanup_package_artefacts t nv =
   log "Cleaning up artefacts of %a" (slog OpamPackage.to_string) nv;
 
   let build_dir = OpamPath.Switch.build t.switch_global.root t.switch nv in
-  if not OpamStateConfig.(!r.keep_build_dir) &&
-     OpamFilename.exists_dir build_dir then
-    OpamFilename.rmdir build_dir;
+  if not OpamStateConfig.(!r.keep_build_dir) then OpamFilename.rmdir build_dir;
+  let remove_dir = OpamPath.Switch.remove t.switch_global.root t.switch nv in
+  if OpamFilename.exists_dir remove_dir then OpamFilename.rmdir remove_dir;
   let name = nv.name in
   let dev_dir =
     OpamPath.Switch.dev_package t.switch_global.root t.switch name
@@ -528,17 +519,19 @@ let sources_needed st g =
       | _ -> assert false)
     g OpamPackage.Set.empty
 
-let remove_package t ?keep_build ?silent ?changes ?force nv =
+let remove_package t ?silent ?changes ?force nv =
   if OpamStateConfig.(!r.fake) || OpamStateConfig.(!r.show) then
     Done (OpamConsole.msg "Would remove: %s.\n" (OpamPackage.to_string nv))
   else
-    remove_package_aux t ?keep_build ?silent ?changes ?force nv
+    remove_package_aux t ?silent ?changes ?force nv
 
 (* Compiles a package.
    Assumes the package has already been downloaded to [source].
 *)
 let build_package t source nv =
-  extract_package t source nv;
+  let build_dir = OpamPath.Switch.build t.switch_global.root t.switch nv in
+  OpamFilename.rmdir build_dir;
+  extract_package t source nv build_dir;
   let opam = OpamSwitchState.opam t nv in
   let commands =
     OpamFile.OPAM.build opam @
@@ -552,11 +545,10 @@ let build_package t source nv =
   in
   let env = OpamTypesBase.env_array (compilation_env t opam) in
   let name = OpamPackage.name_to_string nv in
-  let dir = OpamPath.Switch.build t.switch_global.root t.switch nv in
   let rec run_commands = function
     | (cmd::args)::commands ->
       let text = OpamProcess.make_command_text name ~args cmd in
-      let dir = OpamFilename.Dir.to_string dir in
+      let dir = OpamFilename.Dir.to_string build_dir in
       let cmd, args = cmd_wrapper t opam OpamFile.Config.wrap_build cmd args in
       OpamSystem.make_command ~env ~name ~dir ~text
         ~verbose:(OpamConsole.verbose ()) ~check_existence:false
@@ -570,7 +562,14 @@ let build_package t source nv =
            name (String.concat " " (cmd::args));
          Done (Some (OpamSystem.Process_error result)))
     | []::commands -> run_commands commands
-    | [] -> Done None
+    | [] ->
+      OpamConsole.msg "%s compiled  %s.%s\n"
+          (if not (OpamConsole.utf8 ()) then "->"
+           else OpamActionGraph.
+                  (action_color (`Build ()) (action_strings (`Build ()))))
+          (OpamConsole.colorise `bold name)
+          (OpamPackage.version_to_string nv);
+        Done None
   in
   run_commands commands
 
@@ -633,7 +632,7 @@ let install_package t nv =
     install_job
   @@+ function
   | Some e, changes ->
-    remove_package t ~keep_build:true ~silent:true ~changes nv @@| fun () ->
+    remove_package t ~silent:true ~changes nv @@| fun () ->
     OpamStd.Exn.fatal e;
     Some e
   | None, changes ->
